feat: configure update and storage at/rt tokens
Logic for refresh token and logout - refreshTokens and logout respectively. Each request is now appended with an access_token if available.

access_token - has a short lifetime and is stored in the application state.

refresh_token - is only stored in cookies (httpOnly, secure, sameSite).

Currently the token refresh logic is in (root)/layout.tsx. *Temporary solution.

Created slice to handle user data, tokens only for now.

Some of the logic has been moved from the header component to separate SignInButton and UserMenu components.

As a result, added functionality for users to sign in and out of their account.